+++ Core/Include/Core/Log.h
@@ -6,6 +6,7 @@
 #include <memory>
 #include <shared_mutex>
 #include <string>
+#include <utility>
 
 #include "spdlog/spdlog.h"
 
@@ -50,8 +51,17 @@ class LoggerHandle
     LoggerHandle(const LoggerHandle&) = delete;
     LoggerHandle& operator=(const LoggerHandle&) = delete;
 
-    spdlog::logger* operator->() const noexcept;
     explicit operator bool() const noexcept;
+    std::size_t SinkCount() const noexcept;
+    void Flush() const;
+    void SetLevel(LogLevel level) const;
+
+    template <typename... Args>
+    void Write(spdlog::source_loc location, spdlog::level::level_enum level, spdlog::format_string_t<Args...> format,
+               Args&&... args) const
+    {
+        m_Logger->log(location, level, format, std::forward<Args>(args)...);
+    }
 
   private:
     friend class Log;
@@ -74,60 +84,62 @@ class Log
     static LoggerHandle GetClientLogger();
 
   private:
+    friend class LoggerHandle;
     static spdlog::level::level_enum ToSpdlogLevel(LogLevel level);
 };
 } // namespace Core
 
-#define CORE_DETAIL_LOG(getter, logMacro, ...)                                                                         \
+#define CORE_DETAIL_LOG(getter, logLevel, ...)                                                                         \
     do                                                                                                                 \
     {                                                                                                                  \
         auto coreDetailLoggerHandle = getter;                                                                          \
-        logMacro(coreDetailLoggerHandle, __VA_ARGS__);                                                                 \
+        coreDetailLoggerHandle.Write(::spdlog::source_loc{__FILE__, __LINE__, SPDLOG_FUNCTION}, logLevel,              \
+                                     __VA_ARGS__);                                                                     \
     } while (false)
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_TRACE
-#define CORE_TRACE(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_TRACE, __VA_ARGS__)
-#define CLIENT_TRACE(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_TRACE, __VA_ARGS__)
+#define CORE_TRACE(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::trace, __VA_ARGS__)
+#define CLIENT_TRACE(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::trace, __VA_ARGS__)
 #else
 #define CORE_TRACE(...) (void)0
 #define CLIENT_TRACE(...) (void)0
 #endif
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_DEBUG
-#define CORE_DEBUG(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_DEBUG, __VA_ARGS__)
-#define CLIENT_DEBUG(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_DEBUG, __VA_ARGS__)
+#define CORE_DEBUG(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::debug, __VA_ARGS__)
+#define CLIENT_DEBUG(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::debug, __VA_ARGS__)
 #else
 #define CORE_DEBUG(...) (void)0
 #define CLIENT_DEBUG(...) (void)0
 #endif
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_INFO
-#define CORE_INFO(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_INFO, __VA_ARGS__)
-#define CLIENT_INFO(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_INFO, __VA_ARGS__)
+#define CORE_INFO(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::info, __VA_ARGS__)
+#define CLIENT_INFO(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::info, __VA_ARGS__)
 #else
 #define CORE_INFO(...) (void)0
 #define CLIENT_INFO(...) (void)0
 #endif
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_WARN
-#define CORE_WARN(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_WARN, __VA_ARGS__)
-#define CLIENT_WARN(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_WARN, __VA_ARGS__)
+#define CORE_WARN(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::warn, __VA_ARGS__)
+#define CLIENT_WARN(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::warn, __VA_ARGS__)
 #else
 #define CORE_WARN(...) (void)0
 #define CLIENT_WARN(...) (void)0
 #endif
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_ERROR
-#define CORE_ERROR(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_ERROR, __VA_ARGS__)
-#define CLIENT_ERROR(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_ERROR, __VA_ARGS__)
+#define CORE_ERROR(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::err, __VA_ARGS__)
+#define CLIENT_ERROR(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::err, __VA_ARGS__)
 #else
 #define CORE_ERROR(...) (void)0
 #define CLIENT_ERROR(...) (void)0
 #endif
 
 #if SPDLOG_ACTIVE_LEVEL <= SPDLOG_LEVEL_CRITICAL
-#define CORE_CRITICAL(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), SPDLOG_LOGGER_CRITICAL, __VA_ARGS__)
-#define CLIENT_CRITICAL(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), SPDLOG_LOGGER_CRITICAL, __VA_ARGS__)
+#define CORE_CRITICAL(...) CORE_DETAIL_LOG(::Core::Log::GetCoreLogger(), ::spdlog::level::critical, __VA_ARGS__)
+#define CLIENT_CRITICAL(...) CORE_DETAIL_LOG(::Core::Log::GetClientLogger(), ::spdlog::level::critical, __VA_ARGS__)
 #else
 #define CORE_CRITICAL(...) (void)0
 #define CLIENT_CRITICAL(...) (void)0